MS-34: create modal and styles
