--- src/Components/Postulants/Modal/index.jsx
@@ -1,30 +1,0 @@
-import styles from './modal.module.css';
-
-const Modal = (props) => {
-  if (!props.showModal) {
-    return null;
-  }
-
-  const deletePostulant = () => {
-    props.actionPostulant();
-    props.closeModal();
-  };
-
-  return (
-    <div className={styles.container}>
-      <div className={styles.modal}>
-        <h3>{props.titleText}</h3>
-        <p>{props.warningText}</p>
-        <p>
-          {props.text} {props.firstName} {props.lastName}
-        </p>
-        <div className={styles.buttonContainer}>
-          <button onClick={deletePostulant}>{props.buttonText}</button>
-          <button onClick={() => props.closeModal()}>close</button>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-export default Modal;
