--- src/Components/Postulants/Modal/modal.module.css
@@ -1,35 +1,0 @@
-.container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100vh;
-  background-color: rgba(0, 0, 0, 0.5);
-}
-
-.modal {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  min-width: 400px;
-  min-height: 200px;
-  background-color: lightgray;
-}
-
-
-.modal h3, p {
-  align-self: center;
-}
-
-.modal button {
-  cursor: pointer;
-}
-
-.buttonContainer {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-}
